// packages/schema/persistence.d.ts
export type Projects = ProjectSummary[];
// ...
export interface ProjectList {
  projects: Projects;
}
export interface ProjectSummary {
  name: Name;
  projectId: Projectid;
  revision: Revision;
}